Update test to use floats instead of ints.

+++ test/fusion/test_core.py
@@ -329,7 +329,7 @@ class TestDecorator(unittest.TestCase):
         self._check(actual, expected)
 
     def test_fusing_stencils(self):
-        in_grid = (numpy.random.rand(1024, 1024) * 10).astype(numpy.int32).astype(numpy.float32)
+        in_grid = numpy.random.rand(1024, 1024).astype(numpy.float32) * 10
         kernel1 = Stencil(backend='ocl')
         kernel2 = Stencil(backend='ocl')
         py_kernel1 = Stencil(backend='ocl')
